Add files via upload
•	The main objective of the project is to automate the process of detection of skin disease with the help of Convolutional Neural Network (CNN). 
•	It provides accurate and fast results of classification. This technique reduces human intervention and reduces the error.
•	To provide a web tool for the detection of skin disease.
The dataset that is being used in the project is HAM 10000 (Human Against Machine) which is published by Harvard University consisting of 10015 images of seven different categories.
They are Actinic keratosis, Basal cell carcinoma, Benign keratosis, Dermatofibroma, Melanoma, Melanocytic nevi, Vascular lesions
Colour of the input image is very important in the detection of skin disease. 
The steps involve are:
•	The colour image of the skin lesion from the dataset is taken as the input. 
•	These images are first pre-processed. 
•	The dataset is then divided into training set and validation set. 
•	Then, the model is built.
•	 Then, we evaluate the model 
•	Build web tool “Derm Lens” to detect skin diseases
The proposed CNN model, gives 91% training accuracy and 71% testing accuracy.

This project presented a technique for the detection of skin disease where pre-processing is done by using morphological Black-Hat transform and Convolutional Neural Network (CNN) is built for feature extraction and classification. CNNs stand to be robust and efficient methodology for solving a classification problem as they provide good accuracy. 

	With the advancement in technology every task can be automated. This automation is very essential for the man kind to solve various issues.